requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
yfinance
pandas
aiohttp
jinja2